+++ examples/plot_auto_repair.py
@@ -136,5 +136,4 @@ plt.tight_layout()
 # To top things up, we can also visualize the bad sensors for each trial using
 # a heatmap.
 
-fig = ar.get_reject_log(epochs['Auditory/Left']).plot('vertical')
-fig.gca().set_xticks(range(0, len(epochs.ch_names), 10))
+fig = ar.get_reject_log(epochs['Auditory/Left']).plot()