+++ doc/whats_new.rst
@@ -17,6 +17,8 @@ Changelog
 
 - Make plotting of reject log more flexible by allowing to change the orientation of plot, option to supress the immediate display of the plot and returning the plot figure for further modification or saving in :meth:`autoreject.RejectLog.plot`, by `Marcin Koculak`_ in `#152 <https://github.com/autoreject/autoreject/pull/152>`_
 
+- Add `show_names` option to `:func:autoreject.RejectLog.plot` by `Mainak Jas`_ in `#322 <https://github.com/autoreject/autoreject/pull/322>`_
+
 Bug
 ~~~